farmer2market/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cachetools==6.2.6
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.6
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
deprecation==2.1.0
dj-database-url==3.1.2
Django==5.2.12
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
fsspec==2026.2.0
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
msgpack==1.1.2
multidict==6.7.1
packaging==26.0
pillow==12.1.1
postgrest==2.28.3
propcache==0.4.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyOpenSSL==26.1.0
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
realtime==2.28.3
redis==7.4.0
requests==2.32.5
rich==14.3.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
urllib3==2.6.3
websockets==15.0.1
whitenoise==6.12.0
yarl==1.23.0
zope.interface==8.4
zstandard==0.25.0
